app-admin/profile-cleaner: Update to 2.37
app-admin/profile-cleaner: Update from 2.36 to 2.37, the latest version
not available in Portage tree yet. The ebuild is based on
profile-cleaner-2.36.ebuild in Portage tree, with an extra command added
to src_install() that removes a symlink to the man page, which will
typically be broken by the recompression of man pages in Portage.